SMB-73: Déplacement des méthodes calculant l'autocorrélation
Elles sont maintenant situées dans RegressionLogistique-as.cpp.
